# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/notifications": {
      "get": {
        "operationId": "NotificationsController_list",
        "parameters": [
          {
            "name": "unreadOnly",
            "required": false,
            "in": "query",
            "description": "When true, only unread notifications are returned",
            "schema": {
              "default": false,
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "th_session": []
          }
        ],
        "summary": "Lists the signed-in user's own notifications, newest first",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/unread-count": {
      "get": {
        "operationId": "NotificationsController_unreadCount",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnreadCountResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "th_session": []
          }
        ],
        "summary": "The signed-in user's unread notification count",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/{id}/read": {
      "post": {
        "operationId": "NotificationsController_markRead",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "th_session": []
          }
        ],
        "summary": "Marks one of the signed-in user's own notifications as read",
        "tags": [
          "notifications"
        ]
      }
    },
    "/notifications/read-all": {
      "post": {
        "operationId": "NotificationsController_markAllRead",
        "parameters": [],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "th_session": []
          }
        ],
        "summary": "Marks all of the signed-in user's notifications as read",
        "tags": [
          "notifications"
        ]
      }
    }
  },
  "info": {
    "title": "Telehealth API",
    "description": "REST API for the telehealth prototype",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "th_session",
        "description": "Opaque session token set by POST /auth/login or a registration endpoint."
      }
    },
    "schemas": {
...
      "NotificationResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "APPOINTMENT_BOOKED",
              "BOOKING_CONFIRMED",
              "APPOINTMENT_RESCHEDULED",
              "RESCHEDULE_CONFIRMED",
              "APPOINTMENT_CANCELLED",
              "REMINDER_24H",
              "REMINDER_1H"
            ]
          },
          "title": {
            "type": "string"
          },
          "body": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "additionalProperties": true,
            "nullable": true,
            "description": "Structured event data (startsAt, previousStartsAt, counterpartName, reason) for the viewer to format."
          },
          "link": {
            "type": "string",
            "nullable": true,
            "description": "A link to the appointment in the recipient's role area"
          },
          "appointmentId": {
            "type": "string",
            "nullable": true
          },
          "readAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "type",
          "title",
          "body",
          "data",
          "link",
          "appointmentId",
          "readAt",
          "createdAt"
        ]
      },
      "NotificationListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NotificationResponseDto"
            }
          },
          "total": {
            "type": "number"
          },
          "page": {
            "type": "number"
          },
          "pageSize": {
            "type": "number"
          },
          "unreadCount": {
            "type": "number"
          }
        },
        "required": [
          "items",
          "total",
          "page",
          "pageSize",
          "unreadCount"
        ]
      },
      "UnreadCountResponseDto": {
        "type": "object",
        "properties": {
          "unreadCount": {
            "type": "number"
          }
        },
        "required": [
          "unreadCount"
        ]
      }
    }
  }
}